J-286: +ready-tabs-navigation
