+++ src/App.tsx
@@ -7,16 +7,13 @@ import { NavigationContainer } from '@react-navigation/native'
 import { Navigation } from './navigation'
 import { store } from './store'
 import styles from 'src/styles/globalStyles'
-import { TabsNavigator } from './components'
-
-const isSigned = true
 
 const App = (): React.ReactElement => {
   return (
     <NavigationContainer>
       <Provider store={store}>
         <SafeAreaView style={styles.container}>
-          {isSigned ? <TabsNavigator /> : <Navigation />}
+          <Navigation />
           <Toast ref={(ref) => Toast.setRef(ref)} />
         </SafeAreaView>
       </Provider>
