+++ src/common/enums/navigation/navigationScreen.ts
@@ -1,6 +1,7 @@
 enum NavigationScreen {
   HOME = 'Home',
-  SIGN_IN = 'SignIn'
+  SIGN_IN = 'SignIn',
+  TABS_NAV = 'TabsNav'
 }
 
 export { NavigationScreen }
