+++ src/components/index.ts
@@ -2,4 +2,3 @@ export { default as Button, ButtonType } from './Button'
 export { default as Input } from './Input'
 export { default as Link } from './Link'
 export { default as Heading, HeadingType } from './Heading'
-export { default as TabsNavigator } from './navigationTab'
