+++ src/navigation/navigation.tsx
@@ -1,11 +1,18 @@
 import React from 'react'
+import { View, Text } from 'react-native'
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
+import { Home, PodcastPage, SignIn } from 'src/screens'
 import { createStackNavigator } from '@react-navigation/stack'
-import { SignIn, Home } from 'src/screens'
 import { NavigationScreen } from 'src/common/enums'
+import HomeIcon from 'src/assets/images/home.svg'
+import Play from 'src/assets/images/playNavigation.svg'
+import User from 'src/assets/images/meIcon.svg'
+import styles from './styles'
 
 const Stack = createStackNavigator()
+const Tab = createBottomTabNavigator()
 
-const Navigation: React.FC = () => {
+const Nav: React.FC = () => {
   return (
     <Stack.Navigator
       initialRouteName={NavigationScreen.SIGN_IN}
@@ -15,9 +22,103 @@ const Navigation: React.FC = () => {
       }}
     >
       <Stack.Screen name={NavigationScreen.SIGN_IN} component={SignIn} />
-      <Stack.Screen name={NavigationScreen.HOME} component={Home} />
+      <Stack.Screen name={NavigationScreen.TABS_NAV} component={navigation} />
     </Stack.Navigator>
   )
 }
 
-export default Navigation
+const navigation: React.FC = () => {
+  return (
+    <Tab.Navigator
+      initialRouteName={NavigationScreen.HOME}
+      screenOptions={{
+        headerShown: false,
+        tabBarShowLabel: false,
+        tabBarInactiveTintColor: '#fff',
+        tabBarActiveTintColor: '#fff',
+        tabBarStyle: {
+          backgroundColor: 'rgba(25,25,25,0.95)',
+          height: 70,
+          borderRadius: 10,
+          marginBottom: 10,
+          marginHorizontal: 10,
+          elevation: 5
+        }
+      }}
+    >
+      <Tab.Screen
+        name="Home"
+        component={Home}
+        options={{
+          tabBarIcon: ({ focused }) => {
+            if (focused) {
+              return (
+                <View style={styles.viewContainer}>
+                  <HomeIcon width={17} />
+                  <Text style={styles.textNavigation}>Home</Text>
+                </View>
+              )
+            } else {
+              return (
+                <View style={styles.viewContainer}>
+                  <HomeIcon width={17} opacity={0.2} />
+                  <Text style={styles.textNavigationWithOpacity}>Home</Text>
+                </View>
+              )
+            }
+          }
+        }}
+      />
+      <Tab.Screen
+        name="Listening"
+        component={PodcastPage}
+        options={{
+          tabBarIcon: ({ focused }) => {
+            if (focused) {
+              return (
+                <View style={styles.viewContainer}>
+                  <Play width={17} />
+                  <Text style={styles.textNavigation}>Listening</Text>
+                </View>
+              )
+            } else {
+              return (
+                <View style={styles.viewContainer}>
+                  <Play width={17} opacity={0.2} />
+                  <Text style={styles.textNavigationWithOpacity}>
+                    Listening
+                  </Text>
+                </View>
+              )
+            }
+          }
+        }}
+      />
+      <Tab.Screen
+        name="MyProfile"
+        component={SignIn}
+        options={{
+          tabBarIcon: ({ focused }) => {
+            if (focused) {
+              return (
+                <View style={styles.viewContainer}>
+                  <User width={17} />
+                  <Text style={styles.textNavigation}>Me</Text>
+                </View>
+              )
+            } else {
+              return (
+                <View style={styles.viewContainer}>
+                  <User width={17} opacity={0.2} />
+                  <Text style={styles.textNavigationWithOpacity}>Me</Text>
+                </View>
+              )
+            }
+          }
+        }}
+      />
+    </Tab.Navigator>
+  )
+}
+
+export default Nav
