+++ src/navigation/styles.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  viewContainer: { alignItems: 'center', flexDirection: 'row' },
+  textNavigation: { color: '#fff', marginLeft: 10, fontSize: 12 },
+  textNavigationWithOpacity: {
+    color: '#ffffff66',
+    marginLeft: 10,
+    fontSize: 12
+  }
+})
+
+export default styles
